Moving the chooseHeadCase function in the Transformations file + some comments
